+++ pkg/validator/handler.go
@@ -86,6 +86,8 @@ func errorResponse(uid types.UID, err error) *v1beta1.AdmissionResponse {
 		Allowed: false,
 		UID:     uid,
 		Result: &metav1.Status{
+			Reason:  metav1.StatusReasonBadRequest,
+			Code:    http.StatusBadRequest,
 			Message: err.Error(),
 		},
 	}