+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Set `400` status code if a request is invalid.
+
 ## [1.9.1] - 2020-10-21
 
 ### Fixed