+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.10.0] - 2020-10-23
+
 ### Added
 
 - AzureCluster, AzureMachine, AzureMachinePool, Cluster and MachinePools CR's
@@ -109,7 +111,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First release.
 
-[Unreleased]: https://github.com/giantswarm/azure-admission-controller/compare/v1.9.1...HEAD
+[Unreleased]: https://github.com/giantswarm/azure-admission-controller/compare/v1.10.0...HEAD
+[1.10.0]: https://github.com/giantswarm/azure-admission-controller/compare/v1.9.1...v1.10.0
 [1.9.1]: https://github.com/giantswarm/azure-admission-controller/compare/v1.9.0...v1.9.1
 [1.9.0]: https://github.com/giantswarm/azure-admission-controller/compare/v1.8.0...v1.9.0
 [1.8.0]: https://github.com/giantswarm/azure-admission-controller/compare/v1.7.0...v1.8.0